backend/requirements.txt
alembic==1.18.4
annotated-types==0.7.0
blinker==1.9.0
click==8.3.1
colorama==0.4.6
dnspython==2.8.0
email-validator==2.1.0.post1
Flask==3.0.0
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
greenlet==3.3.1
gunicorn==21.2.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
packaging==26.0
psycopg2-binary==2.9.9
pydantic==2.5.2
pydantic_core==2.14.5
PyJWT==2.11.0
python-dotenv==1.0.0
SQLAlchemy==2.0.46
typing_extensions==4.15.0
Werkzeug==3.1.5
